pyproject.toml
[project]
name = "torchtmpl"
version = "0.0.1"

dependencies = [
    "PyYAML>=6.0",
    "torch>=1.13.1",
    "torchvision>=0.14.1",
    "torchinfo>=1.7.2",
    "wandb>=0.13.11",
    "scipy>=1.10.1",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.1",
    "pandas"
]

[options]
package_dir = "torchtmpl"
